Fix composeMiddleware and move back into separate module
Discovered a subtle and confounding bug with composeMiddleware where
it only works if dispatch is the last argument. It did not combine
multiple middlewares into a single middlewares as advertised; it
only combined multiple middlewares with dispatch to create a new
dispatch. This didn't come up earlier because the codebase never
happened to use it in the former way.

This commit fixes the issue and adds a test for it.


Former-commit-id: 9a1a0cd1349563c8e4c4fa6adbc38247146098fb
Former-commit-id: 88c136ac33016d5945deb58e9aca5e2bab1ef39b