Add missing reducer composition example to docs
The code for the first implementation of `todos` as an example of reducer composition is actually the same as the previous example, causing the following paragraph ("This is called *reducer composition*, and it’s the fundamental pattern of building Redux apps") to make no sense. 

This change adds the `todos` function from the following example to make this one more clear.

There are other ways the composition could be done, certainly, and I'm not sure that the one I chose is the most iconic, so feel free to suggest alternatives. :-)

Former-commit-id: 90ef233b6e7b934d01758c8ac44fb8008c077a02
Former-commit-id: ee8ea865f834143bc3b98c88f3425b9220511f85