Add a Gitbook plugin to allow correct loading of doc page hash URLs.
Per #345, navigating to a docs page URL that includes a hash anchor
causes the page to snap back up to the top.  This is due to Gitbook
showing the sidebar.  The "anker-enable" plugin hacks around that by
setting a short timeout on docs page load, and then changing the page
location back to the hash URL.  This appears to work around the issue.


Former-commit-id: 5028fae9d2b73049fea8a3aeff4f5dd04bf93dd2
Former-commit-id: e7277e7602c2a9ea6d68540abaf4f23b06955f9c