document assertReducerSanity behavior (#1633)
It is not clear from the documentation that it is insufficient to
merely avoid ever passing in `undefined` as state by using initial
state. `assertReducerSanity` will test that your reducer does the
right thing when passed `undefined`, even if you never intend for it
to be called this way.

This documentation change makes that behavior clearer, in the hopes of
saving new users from needless frustration.

Former-commit-id: 31f0305fe4dfc3b095a138cedea01a91d7d266c5
Former-commit-id: 3ecf9519f0b8ac585dfd79fc1195928ec124704f