Annotate `action.js` headings (#2571)
As it stands, there are multiple headings entitled `action.js`. On the official docs page, it offers a anchor to each heading, but this doesn't work because it will just match the `id` of the first `#### action.js` heading. This change adds some context, which will—in turn—make each heading unique.

Former-commit-id: 5dc65541d1347166d634727cd5f4dc12f627f8af
Former-commit-id: cb8deab72b6cbb352c23d985c773359726dca747