Remove outdated Link, Add Redux Book (#3232)
Just went through the Redux resources page and saw an outdated link from my website. It shows how to use normalizr, but I feel guilty because I really need to update it. 

A better resource for learning Redux in React would be Taming the State in React though :) Would love to get it into the books section.

Former-commit-id: ad4f06f1f5376e8d3c8efe4a049988e1a197faf7
Former-commit-id: 3a37ad353443384c2426fc816e694243dc2f8171