Update UsingObjectSpreadOperator.md (#3367)
I think this warning on using the Object spread operator to clone objects is important. While, good style would encourage the use of flat objects in the reducer, not everyone is going to do that. Using the spread syntax in a reducer on a multidimensional data structure is opening up the door to state mutation.

Former-commit-id: 4430ae546aced1d478311cdafd01365b948b1d81
Former-commit-id: 63bf180d00b5e544dfc8e9961a2ebff8d52ef728